+++ app/lti/services/ags/line_items.py
@@ -1,0 +1,26 @@
+from django.conf import settings
+
+domain = settings.LTI_PLATFORM_DOMAIN
+
+
+class AGSLineItemService:
+
+    @staticmethod
+    def create():
+        pass
+
+    @staticmethod
+    def update():
+        pass
+
+    @staticmethod
+    def get():
+        pass
+
+    @staticmethod
+    def list():
+        pass
+
+    @staticmethod
+    def delete():
+        pass
