+++ app/lti/services/ags/request.py
@@ -1,0 +1,13 @@
+class AGSRequest:
+
+    def get(self):
+        pass
+
+    def post(self):
+        pass
+
+    def put(self):
+        pass
+
+    def delete(self):
+        pass
