+++ app/lti/services/ags/scores.py
@@ -1,0 +1,9 @@
+class AGSScoresService:
+
+    @staticmethod
+    def results():
+        pass
+
+    @staticmethod
+    def progress():
+        pass
