+++ app/materia/settings/lti.py
@@ -29,6 +29,8 @@ LTI_URL_CONFIGS = {
     "platform_iss": os.environ.get("PLATFORM_ISS"),
 }
 
-CANVAS_OAUTH_CLIENT_ID = os.environ.get("CANVAS_OAUTH_CLIENT_ID")
-CANVAS_OAUTH_CLIENT_SECRET = os.environ.get("CANVAS_OAUTH_CLIENT_SECRET")
-CANVAS_OAUTH_CANVAS_DOMAIN = os.environ.get("CANVAS_DOMAIN")
+LTI_PLATFORM_DOMAIN = os.environ.get("CANVAS_DOMAIN")
+
+# CANVAS_OAUTH_CLIENT_ID = os.environ.get("CANVAS_OAUTH_CLIENT_ID")
+# CANVAS_OAUTH_CLIENT_SECRET = os.environ.get("CANVAS_OAUTH_CLIENT_SECRET")
+# CANVAS_OAUTH_CANVAS_DOMAIN = os.environ.get("CANVAS_DOMAIN")
